requirements.txt
amqp==2.6.1
appdirs==1.4.4
billiard==3.5.0.5
black==21.4b2
celery==4.2.1
certifi==2018.4.16
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
dataclasses==0.8
Django==2.1
idna==2.6
importlib-metadata==4.0.1
install==1.3.4
jsonschema==2.6.0
kombu==4.6.11
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pycparser==2.20
pymodm==0.4.0
pymongo==3.11.3
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
redis==3.5.3
regex==2021.4.4
requests==2.18.4
six==1.11.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==1.35
urllib3==1.22
vine==1.3.0
zipp==3.4.1
